+++ bot/register.py
@@ -1,4 +1,16 @@
+"""
+Register a user interaction
+"""
+import json
+import logging
+import os
+import subprocess
+import numpy
+import requests
+
 from checks import CheckResident, CheckCondo
+from db.schema import create_resident
+from notify import NotifyAdmin
 from python_speech_features import mfcc
 from scipy.io.wavfile import read
 from settings import LOG_NAME
@@ -7,20 +19,21 @@ from settings import PATH
 from telegram import KeyboardButton, ReplyKeyboardMarkup
 from telegram.ext import ConversationHandler
 from validator import ValidateForm
-import json
-import logging
-import numpy
-import os
-import requests
-import subprocess
+
 
 logger = logging.getLogger(LOG_NAME)
 
 chat = {}
 
 class Register:
+    """
+    Register a resident
+    """
 
     def index(update, context):
+        """
+        Start the conversation
+        """
         logger.info("Introducing registration session")
         chat_id = update.message.chat_id
 
@@ -35,6 +48,9 @@ class Register:
         return NAME
 
     def name(update, context):
+        """
+        Get name of a resident
+        """
         chat_id = update.message.chat_id
         name = update.message.text
 
@@ -45,8 +61,12 @@ class Register:
         logger.debug(f"'name': '{chat[chat_id]['name']}'")
 
         contact_keyboard = KeyboardButton('Enviar meu número de telefone', request_contact=True)
-        custom_keyboard = [[ contact_keyboard ]]
-        reply_markup = ReplyKeyboardMarkup(custom_keyboard, one_time_keyboard=True, resize_keyboard=True)
+        custom_keyboard = [[contact_keyboard]]
+        reply_markup = ReplyKeyboardMarkup(
+            custom_keyboard,
+            one_time_keyboard=True,
+            resize_keyboard=True
+        )
 
         update.message.reply_text('Telefone:', reply_markup=reply_markup)
         logger.info("Asking for phone")
@@ -54,6 +74,9 @@ class Register:
         return PHONE
 
     def phone(update, context):
+        """
+        Get phone information
+        """
         chat_id = update.message.chat_id
         phone = update.message.text
         contact = update.effective_message.contact
@@ -72,6 +95,10 @@ class Register:
         return EMAIL
 
     def email(update, context):
+        """
+        Get email information
+        """
+
         chat_id = update.message.chat_id
         email = update.message.text
 
@@ -96,6 +123,10 @@ class Register:
         return CPF
 
     def cpf(update, context):
+        """
+        Get cpf information
+        """
+
         chat_id = update.message.chat_id
         cpf = update.message.text
 
@@ -124,6 +155,10 @@ class Register:
 
 
     def block(update, context):
+        """
+        Get block information
+        """
+
         chat_id = update.message.chat_id
         block = update.message.text
 
@@ -148,6 +183,10 @@ class Register:
         return APARTMENT
 
     def apartment(update, context):
+        """
+        Get apartment information
+        """
+
         chat_id = update.message.chat_id
         apartment = update.message.text
 
@@ -167,13 +206,19 @@ class Register:
         logger.debug("Existing apartment - proceed")
 
         update.message.reply_text(
-            'Vamos agora cadastrar a sua voz! Grave uma breve mensagem de voz dizendo "Juro que sou eu"')
+            'Vamos agora cadastrar a sua voz! Grave uma breve mensagem de' +
+            ' voz dizendo "Juro que sou eu"'
+        )
 
         logger.info("Requesting voice audio")
 
         return VOICE_REGISTER
 
     def voice_register(update, context):
+        """
+        Get voice from the user
+        """
+
         chat_id = update.message.chat_id
         voice_register = update.message.voice
 
@@ -199,20 +244,30 @@ class Register:
         chat[chat_id]['voice_reg'] = None
         chat[chat_id]['voice_mfcc'] = mfcc_data
         logger.debug(f"'voice_reg': '{chat[chat_id]['voice_reg']}'")
-        logger.debug(f"'voice_mfcc': '{chat[chat_id]['voice_mfcc'][:1]}...{chat[chat_id]['voice_mfcc'][-1:]}'")
+        logger.debug(
+            f"'voice_mfcc': '{chat[chat_id]['voice_mfcc'][:1]}.."+
+            ".{chat[chat_id]['voice_mfcc'][-1:]}'"
+        )
 
         # Repeat and confirm buttons
         repeat_keyboard = KeyboardButton('Repetir')
         confirm_keyboard = KeyboardButton('Confirmar')
-        keyboard = [[repeat_keyboard],[confirm_keyboard]]
+        keyboard = [[repeat_keyboard], [confirm_keyboard]]
         choice = ReplyKeyboardMarkup(keyboard, resize_keyboard=True, one_time_keyboard=True)
-        update.message.reply_text('Escute o seu áudio e confirme se está com boa qualidade', reply_markup = choice)
+        update.message.reply_text(
+            'Escute o seu áudio e confirme se está com boa qualidade',
+            reply_markup=choice
+        )
 
         logger.info("Asking to confirm or repeat voice audio")
 
         return REPEAT_VOICE
 
     def repeat_voice(update, context):
+        """
+        Repeate voice interaction
+        """
+
         chat_id = update.message.chat_id
         choice = update.message.text
 
@@ -223,22 +278,33 @@ class Register:
 
         logger.debug("Confirming voice audio")
 
+        logger.debug(f"data['{chat_id}']: {chat[chat_id]}")
         response = Register.register_resident(chat_id)
 
         if(response.status_code == 200 and 'errors' not in response.json().keys()):
-            logger.info("resident registered in database")
-            update.message.reply_text('Morador cadastrado no sistema!')
+            logger.info("Resident registered in database")
+            update.message.reply_text("Cadastro feito!\nAguarde aprovação dos administradores.")
+
+            create_resident(
+                    cpf=chat[chat_id]['cpf'],
+                    block=chat[chat_id]['block'],
+                    apartment=chat[chat_id]['apartment'],
+                    chat_id=chat_id
+                    )
+            NotifyAdmin.send_message(context, chat[chat_id])
         else:
             logger.error("Registration failed")
             update.message.reply_text('Falha ao cadastrar no sistema!')
-
-        logger.debug(f"data['{chat_id}']: {chat[chat_id]}")
 
         chat[chat_id] = {}
 
         return ConversationHandler.END
 
     def end(update, context):
+        """
+        Cancel interaction
+        """
+
         logger.info("Canceling registration")
         chat_id = update.message.chat_id
 
@@ -251,6 +317,10 @@ class Register:
 
 
     def register_resident(chat_id):
+        """
+        Register a resident
+        """
+
         logger.info("Registering resident")
         query = """
         mutation createResident(
@@ -305,4 +375,3 @@ class Register:
         logger.debug(f"Response: {response.json()}")
 
         return response
-